Pull date logic out of config.go
