+++ internal/config/config.go
@@ -6,7 +6,6 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
-	"time"
 	"yaylog/internal/consts"
 
 	"github.com/spf13/pflag"
@@ -15,12 +14,6 @@ import (
 type SizeFilter struct {
 	StartSize    int64
 	EndSize      int64
-	IsExactMatch bool
-}
-
-type DateFilter struct {
-	StartDate    time.Time
-	EndDate      time.Time
 	IsExactMatch bool
 }
 
@@ -130,58 +123,6 @@ func ParseFlags(args []string) (Config, error) {
 		SortBy:            sortBy,
 		ColumnNames:       columnsParsed,
 	}, nil
-}
-
-func parseDateFilter(dateFilterInput string) (DateFilter, error) {
-	if dateFilterInput == "" {
-		return DateFilter{}, nil
-	}
-
-	if dateFilterInput == ":" {
-		return DateFilter{}, fmt.Errorf("invalid date filter: ':' must be accompanied by a date")
-	}
-
-	pattern := `^(\d{4}-\d{2}-\d{2})?(?::(\d{4}-\d{2}-\d{2})?)?$`
-	re := regexp.MustCompile(pattern)
-	matches := re.FindStringSubmatch(dateFilterInput)
-	isExactMatch := !strings.Contains(dateFilterInput, ":")
-
-	if matches == nil {
-		return DateFilter{}, fmt.Errorf("invalid date filter format: %q", dateFilterInput)
-	}
-
-	startDate, err := parseDateMatch(matches[1], time.Time{})
-	if err != nil {
-		return DateFilter{}, err
-	}
-
-	endDate, err := parseDateMatch(matches[2], time.Now())
-	if err != nil {
-		return DateFilter{}, err
-	}
-
-	return DateFilter{
-		startDate,
-		endDate,
-		isExactMatch,
-	}, nil
-}
-
-func parseDateMatch(dateInput string, defaultDate time.Time) (time.Time, error) {
-	if dateInput == "" {
-		return defaultDate, nil
-	}
-
-	return parseValidDate(dateInput)
-}
-
-func parseValidDate(dateInput string) (time.Time, error) {
-	parsedDate, err := time.Parse(consts.DateOnlyFormat, dateInput)
-	if err != nil {
-		return time.Time{}, err
-	}
-
-	return parsedDate, nil
 }
 
 func parseSizeFilter(sizeFilterInput string) (SizeFilter, error) {
